refactor(vats): factor out makeNameAdmins
 - spell binaryVoteCounter consistently
